implemented Router and fixed file structure
